requirements.txt
bs4
pycrypto
configparser
mysql-connector-python
